Refactor query hooks and component imports for improved modularity
- Removed several unused hooks from query.ts to streamline the codebase.
- Updated component imports to utilize local hooks instead of centralized query functions, enhancing code organization and maintainability.
- Adjusted imports in ProjectLibrary, ProjectLibraryAspect, ProjectLibraryInsight, and ProjectLibraryView components for better modularity.
